+++ spring-boot-starter-core/src/main/java/com/codeartist/component/core/support/cache/annotation/CacheLock.java
@@ -29,7 +29,7 @@ public @interface CacheLock {
     /**
      * 缓存过期时间
      */
-    long timeout();
+    int timeout();
 
     /**
      * 缓存过期时间单位
